# Allocating response buffer based on body size and on header size when…