--- a/CMakeLists.txt
+++ b/CMakeLists.txt
@@ -63,8 +63,7 @@ add_dependencies(hello_world haywire)
 target_link_libraries (hello_world
     ${haywire_location}
     ${CMAKE_SOURCE_DIR}/lib/libuv/.libs/libuv.a
-    ${CMAKE_THREAD_LIBS_INIT}
-    -lm)
+    ${CMAKE_THREAD_LIBS_INIT})
 
 if (${CMAKE_SYSTEM_NAME} MATCHES "Linux")
     target_link_libraries (hello_world rt)