… required and not just statically. Fixes #54.

By default, it will allocate 1K for the headers, but it will reallocate as necessary if the headers take more than 1k.

Since dynamic allocation only kicks in if required, performance is equivalent with the previous version, as the only difference is allocating enough memory to hold the entire body.

### Performance of previous version
```
[ec2-user@ip-172-31-31-0 wrk]$ ./wrk --latency -d 10s -t 8 -c 128 http://172.31.16.12:8000
Running 10s test @ http://172.31.16.12:8000
  8 threads and 128 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   360.21us  305.93us  21.32ms   94.34%
    Req/Sec    47.47k     3.76k   54.69k    53.47%
  Latency Distribution
     50%  294.00us
     75%  369.00us
     90%  506.00us
     99%    1.47ms
  3815801 requests in 10.10s, 574.97MB read
Requests/sec: 377799.31
Transfer/sec:     56.93MB
```

### Performance of this commit
```
[ec2-user@ip-172-31-31-0 wrk]$ ./wrk --latency -d 10s -t 8 -c 128 http://172.31.16.12:8000
Running 10s test @ http://172.31.16.12:8000
  8 threads and 128 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   367.45us  501.67us  40.67ms   96.27%
    Req/Sec    47.62k     2.41k   54.16k    85.52%
  Latency Distribution
     50%  292.00us
     75%  364.00us
     90%  494.00us
     99%    1.64ms
  3827749 requests in 10.10s, 576.45MB read
Requests/sec: 379006.36
Transfer/sec:     57.08MB
```

Also, just proof that this fixes #54:
```
[ec2-user@ip-172-31-31-0 wrk]$ curl http://172.31.16.12:8000
CHAPTER I. Down the Rabbit-Hole  Alice was beginning to get very tired of sitting by her sister on the bank, and of having nothing to do: once or twice she had peeped into the book her sister was reading, but it had no pictures or conversations in it, <and what is the use of a book,> thought Alice <without pictures or conversations?> So she was considering in her own mind (as well as she could, for the hot day made her feel very sleepy and stupid), whether the pleasure of making a daisy-chain would be worth the trouble of getting up and picking the daisies, when suddenly a White Rabbit with pink eyes ran close by her. There was nothing so very remarkable in that; nor did Alice think it so very much out of the way to hear the Rabbit say to itself, <Oh dear! Oh dear! I shall be late!> (when she thought it over afterwards, it occurred to her that she ought to have wondered at this, but at the time it all seemed quite natural); but when the Rabbit actually took a watch out of its waistcoat-pocket, and looked at it, and then hurried on, Alice started to her feet, for it flashed across her mind that she had never before seen a rabbit with either a waistcoat-pocket, or a watch to take out of it, and burning with curiosity, she ran across the field after it, and fortunately was just in time to see it pop down a large rabbit-hole under the hedge. In another moment down went Alice after it, never once considering how in the world she was to get out again. The rabbit-hole went straight on like a tunnel for some way, and then dipped suddenly down, so suddenly that Alice had not a moment to think about stopping herself before she found herself falling down a very deep well. Either the well was very deep, or she fell very slowly, for she had plenty of time as she went down to look about her and to wonder what was going to happen next. First, she tried to look down and make out what she was coming to, but it was too dark to see anything; then she looked at the sides of the well, and noticed that they were filled with cupboards......
```