@kellabyte 2015-12-02
Oh my god you're the best. I've been wanting to get this complete for ages! Thank you so very much to contributing to Haywire. I really do appreciate the help :) I'm going to try and review this in the next couple days and test it and merge it!