@nmdguerreiro 2015-12-03
Got rid of the dependency on the math lib. It's faster to count the number of characters by diving iteratively than using log10.