@kellabyte 2015-12-07
I just did benchmarks in my benchmark environment and it does look like we lost some performance in this implementation. We are losing around `500,000 to 700,000 requests/second`. I ran the benchmark enough times to make sure this wasn't benchmark/environment blips and anomalies. I wonder if it's the `num_chars` or `calloc` calls?

### Before
```
../bin/wrk/wrk --script ./pipelined_get.lua --latency -d 30s -t 40 -c 512 http://server:8000/foo/bar -- 32

Running 30s test @ http://server:8000/foo/bar
  40 threads and 512 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.49ms    3.08ms 145.65ms   88.15%
    Req/Sec   226.94k    43.37k  438.40k    85.31%
  Latency Distribution
     50%    0.88ms
     75%    1.50ms
     90%    6.97ms
     99%    0.00us
  271130610 requests in 30.10s, 39.89GB read
Requests/sec: 9008388.12
Transfer/sec:      1.33GB
```


### After
```
../bin/wrk/wrk --script ./pipelined_get.lua --latency -d 30s -t 40 -c 512 http://server:8000/foo/bar -- 32

Running 30s test @ http://server:8000/foo/bar
  40 threads and 512 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.00ms    3.35ms  97.72ms   85.18%
    Req/Sec   210.50k    58.23k  422.88k    72.83%
  Latency Distribution
     50%    0.98ms
     75%    1.97ms
     90%    9.33ms
     99%    0.00us
  251246262 requests in 30.10s, 36.97GB read
Requests/sec: 8347135.09
Transfer/sec:      1.23GB
```