@nmdguerreiro 2015-12-07
I'll have a look. I think it must be `calloc` call. I'll try `malloc` and null terminating the string, as we don't really need the memory block to be set to 0s. I'll also make sure I run the benchmarks before PR from now on. Worth creating a PR guidelines doc (e.g. run benchmarks)?