@kellabyte 2015-12-07
Yeah I just did another test and there's about a 100MB/s difference even if I set `470` bytes as the buffer default size so that when your calculation does its work it results in `512` which should be aligned I still get a performance drop.

### Before
```
../bin/wrk/wrk --script ./pipelined_get.lua --latency -d 60s -t 40 -c 760 http://server:8000 -- 32

Running 1m test @ http://server:8000
  40 threads and 760 connections
^C  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.70ms    2.63ms 202.77ms   89.23%
    Req/Sec   233.03k    41.34k  447.68k    89.26%
  Latency Distribution
     50%    1.18ms
     75%    1.89ms
     90%    5.03ms
     99%    0.00us
  139138752 requests in 15.04s, 20.47GB read
Requests/sec: 9252834.06
Transfer/sec:      1.36GB
```

### After
```
../bin/wrk/wrk --script ./pipelined_get.lua --latency -d 10s -t 40 -c 760 http://server:8000 -- 32

Running 10s test @ http://server:8000
  40 threads and 760 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.32ms    3.98ms 218.89ms   89.95%
    Req/Sec   212.97k    61.57k  454.95k    74.06%
  Latency Distribution
     50%    1.45ms
     75%    2.56ms
     90%    6.34ms
     99%    0.00us
  85016768 requests in 10.10s, 12.51GB read
Requests/sec: 8417482.61
Transfer/sec:      1.24GB
```